review: switch epoch counter to uint16_t and cover the wrap path
Prior commit excluded the wrap branch from coverage because it was
unreachable in tests with a uint32_t (~4B) counter. uint16_t wraps
every 65,536 searches — frequently enough for a real test to drive
the path (the new "search_layer epoch wrap" test runs 70k searches in
~100 ms and verifies result-correctness across the wrap), but rare
enough that the periodic memset stays out of the bench hot loop.

Bench delta vs. uint32_t is within run-to-run noise across all sizes
(Search ef=10/50, VaryingN 1k/10k/50k, Insert 10k). Memory footprint
halves from 4 to 2 bytes per slot.

+++ src/core/hnsw_index.h
@@ -387,19 +387,19 @@ private:
     // it grows to the high-water mark of count_ across them, and the
     // monotonically-increasing epoch ensures stale marks from a prior search
     // (possibly on a different index) are never read as current.
-    static thread_local std::vector<uint32_t> vis;
-    static thread_local uint32_t vis_epoch = 0;
+    static thread_local std::vector<uint16_t> vis;
+    static thread_local uint16_t vis_epoch = 0;
 
     const size_t total = count_.load(std::memory_order_relaxed);
     assert(ep < total);
     if (vis.size() < total) vis.resize(total, 0);
-    if (++vis_epoch == 0) {  // LCOV_EXCL_START
-      // Epoch wrap requires ~4B searches on this thread before reaching the
-      // reset path; excluded from line coverage as it is not reachable in
-      // any realistic test run.
+    if (++vis_epoch == 0) {
+      // Epoch wraps every 65,536 searches per thread; reset slot state so no
+      // stale mark from a prior epoch can be read as current. Exercised by
+      // the "search_layer epoch wrap" test.
       std::fill(vis.begin(), vis.end(), 0);
       vis_epoch = 1;
-    }  // LCOV_EXCL_STOP
+    }
     vis[ep] = vis_epoch;
     std::priority_queue<std::pair<float, size_t>, std::vector<std::pair<float, size_t>>,
                         std::greater<std::pair<float, size_t>>> cands;
